Implement fine-grained extraction of translatable text
Before, we would extract text based on the byte offsets in the
original document. As a consequence of this, the extracted text would
look precisely like the original: the Markdown was copied directly
from the original. In particular, text from a block quote would
contain the leading ‘>’ characters and paragraphs in list items would
contain leading whitespace.

Now, we instead extract text by grouping the Markdown parse events
into those which should be translated and those who should be skipped.
We use this in two ways:

- When extracting messages in ‘mdbook-xgettext’, we turn the
  translatable events back into Markdown. The structure of the
  document (headings, lists, block quotes, …) is no longer present in
  the extracted messages: only the text content itself it extracted.

- When translating, we replace the sequence of translatable events
  with the events from the translation. We do this while leaving the
  structure of the document unchanged.

The result of this is a much more robust system: editing one list item
no longer impacts adjacent list items, moving a paragraph into a block
quote no longer changes the paragraph.

As a side effect of how we turn events into messages, links are now
all expanded. This makes the messages larger, but it removes a common
source of errors where ‘[foo][1]’ would end up pointing to the wrong
location if the reference link was updated.

Part of #19.