Fix grouping of adjacent code blocks
When grouping Markdown events, we have a goal of making the group
“self-contained”. This roughly means that

    extract_events(reconstruct_markdown(group)) == group

To do this, we included the start/end events for a code block.
However, this had the side effect of merging adjacent code blocks.
This is now fixed. Simultaneously, we fix the extraction of paragraphs
to include the start/end events for those too.