Fix table rendering
When extracting events from a line of text, we get

    &[
        (1, Event::Start(Tag::Paragraph)),
        (1, Event::Text("A line of text".into())),
        (1, Event::End(Tag::Paragraph)),
    ],

This is problematic since the `Text` event has been wrapped in a
paragraph. This is mostly harmless in regular text, but it destroys
the formatting of tables since they cannot contain block-level events
such as paragraphs.

This is now fixed by more carefully trimming the events we inject when
translating.