Add support for skipping translation groups
This adds support for adding a comment of the form: `<!--
mdbook-xgettext: skip -->`. This will cause the system to skip the
next message group that would otherwise be translated.

It adds a dependency to the regex crate to match for the comment skip
pattern.

Tests were added, including tests that cover some odd situations
involving inline HTML, which unfortunately appear to be due to
raphlinus/pulldown-cmark#712.