Fix caching of `target` dir for `fuzz` project
Currently caching does not work for `fuzz` job which can
be verified by looking at Github workflow exec outputs.

The main cause of the issue is the config `workspaces: fuzz -> target`.
The fuzz target does not use a proper workspace but a separate
crate. There is [a pending issue](https://github.com/rust-fuzz/cargo-fuzz/issues/345)
discussing that.

The `rust-cache` action is executed in the root of the project and is
not able to locate `fuzz` directory by workspace name.

I decided to list the directories to cache instead of relying
on target dir detection which seems to work.

I also fixed the path `fuzz/corpus` to `i18n-helpers/fuzz/corpus`.

I adjusted the caching condition `save-if` to cache only on `main`
builds. I think this makes sense because we don't need to cache PRs
which might not get merged or get updated multiple times during review.