Add support for rounding line numbers
This adds a new feature which will let us round line number according
to a user-defined granularity.

The idea is that rounding the line numbers will minimize the updates
to the PO files: with a granularity of 10, the line numbers will be
10, 20, … and this should in turn remove some of the churn on the PO
files.

Setting granularity to zero will remove the line numbers completely
from the PO files.

Fixes #127.