Fix flaky codecov reports
Currently implementation of codecov reports sometimes attaches
the analysis to main branch instead of the triggering PR.
The root cause of the issue is a bug in Github which does not
always populate `github.event.workflow_run.pull_requests[0]`.
When it is populated, the current implementation works. When
it is missing, `override_commit` and `override_pr` params
to `codecov-action` become unset.

More information available at
https://github.com/orgs/community/discussions/25220

The workaround is to store both commit SHA and PR number
in the original workflow since those are available there.