Handle errors returned by `cmark_resume_with_options()`
PR https://github.com/Byron/pulldown-cmark-to-cmark/pull/93 in the
cmark_pulldown_to_cmark crate added structure to the error type returned by
`cmark_resume_with_options()`. It also  made it return an error when the
input is an inconsistent stream of `Event`s.

Now, instead of panicking, we can propagate the error to the caller and
report a more specific error.