Fix broken tables whose msgstr is a specific type of HTML
If an msgstr for a table cell contains HTML that could be parsed as an
HTML block if it were put in the top-level, the table is rendered
incorrectly. The table row containing the cell gets interrupted in the
middle and split into two (or more, if there are more such cells)
incomplete rows.

Close #251.