xgettext: Avoid collisions when splitting .pot files (#243)
PR #77 added the first half of a functionality requested in #67, i.e. being
able to split the catalog across a tree of POT files. However, there are
edge cases where the generated files can have identical path relative to
the po/ directory, with the effect of collapsing messages for unrelated
parts or chapters in the same POT fie.

This commit handles deduplication through a "black box" struct
`UniquePathBuilder`, and adds a doctest and a unit test for collision
scenarios. The new version is backward-compatible for cases that
are collision-free in the old version.

Contextually, simplify the recursion logic in `get_subcontent_for_chapter`.